+++ examples/cartridge-simple/deps.sh
@@ -3,6 +3,10 @@
 
 set -e
 
+# App dependencies:
+tarantoolctl rocks make
+sh ./cartridge.post-build
+
 # Test dependencies:
 tarantoolctl rocks install luatest 0.5.6
 tarantoolctl rocks install luacov 0.13.0